pyproject.toml
[tool.poetry]
name = "grpc"
version = "0.1.0"
description = ""
authors = ["kelvin <kelvininfortec@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "1.66.1"
grpcio-tools = "1.66.1"
protobuf = "5.27.2"
python-dotenv = "1.0.1"

[tool.poetry.dev-dependencies]
pre-commit = "3.6.0"
ipdb = "0.13.13"
black = "23.12.0"
pylint = "2.17.7"
isort = "5.13.2"

[tool.pylint.messages_control]
max-line-length = 88
disable = [
    "C0114",
    "C0115",
    "C0116",
    "R0903",
    "E0611",
    "E1120",
    "R0801",
    "R0913",
    "R0914",
    "C0103",
    "E0401",
    "W0622",
    "I1101",
    "R1710",
    "W0718",
    "R0902",
    "W0150",
    "W1203",
    "C0301",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
force_grid_wrap = 0
include_trailing_comma = true
line_length = 87
multi_line_output = 3
use_parentheses = true